#refactor: agregate migrations
Have to delete all migrations and create a new one because the first two migrations were inherited from the template, and they were causing conflicts.

Have to delete all of them and not just the first two, because the employee entity was hardly changed between migrations, and deleted in the last one, and that was the root of the conflicts.